pyproject.toml
[tool.poetry]
name = "household-manager"
version = "0.1.0"
description = "Advanced Family Budget Manager with OCR, ML categorization, and Real-time updates"
authors = ["Developer <dev@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
django-ninja = "^1.1.0"
django-ninja-extra = "^0.21.0"
pydantic = "^2.6.0"
pydantic-settings = "^2.2.0"
psycopg = {extras = ["binary", "pool"], version = "^3.1.18"}
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.3.1"
argon2-cffi = "^23.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
redis = "^5.0.1"
django-redis = "^5.4.0"
celery = {extras = ["redis"], version = "^5.3.6"}
django-celery-beat = "^2.6.0"
django-celery-results = "^2.5.1"
kombu = "^5.3.5"
channels = {extras = ["daphne"], version = "^4.0.0"}
channels-redis = "^4.2.0"
pytesseract = "^0.3.10"
opencv-python-headless = "^4.9.0"
pillow = "^10.2.0"
scikit-learn = "^1.4.0"
numpy = "^1.26.3"
pandas = "^2.2.0"
joblib = "^1.3.2"
python-telegram-bot = "^21.0"
openpyxl = "^3.1.2"
sentry-sdk = {extras = ["django", "celery"], version = "^1.40.6"}
python-json-logger = "^2.0.7"
prometheus-client = "^0.20.0"
django-prometheus = "^2.3.1"
python-decouple = "^3.8"
whitenoise = {extras = ["brotli"], version = "^6.6.0"}
django-extensions = "^3.2.3"
flower = "^2.0.1"
uvicorn = {extras = ["standard"], version = "^0.28.0"}
daphne = "^4.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-django = "^4.8.0"
pytest-asyncio = "^0.23.5"
pytest-cov = "^4.1.0"
black = "^24.2.0"
ruff = "^0.3.0"
mypy = "^1.8.0"
django-stubs = "^4.2.7"
djangorestframework-stubs = "^3.14.5"
celery-types = "^0.22.0"
ipython = "^8.22.2"
pre-commit = "^3.6.2"
debug-toolbar = "^4.3.0"
factory-boy = "^3.3.0"
faker = "^24.1.0"
httpx = "^0.27.0"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | migrations
)/
'''

[tool.ruff]
line-length = 100
select = [
    "E",
    "W",
    "F",
    "I",
    "C",
    "N",
    "UP",
    "B",
    "A",
    "COM",
    "DJ",
    "PIE",
    "T20",
    "SIM",
    "TCH",
]
ignore = ["E501", "B008", "COM812"]
exclude = ["migrations", ".venv"]

[tool.ruff.isort]
known-first-party = ["apps", "household_manager"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]

[tool.mypy]
python_version = "3.11"
plugins = ["mypy_django_plugin.main"]
ignore_missing_imports = true
strict = false
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true

[tool.django-stubs]
django_settings_module = "household_manager.settings"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "household_manager.settings"
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --tb=short -p no:warnings"
filterwarnings = ["ignore::DeprecationWarning", "ignore::PendingDeprecationWarning"]

[tool.coverage.run]
source = ["apps"]
omit = ["*/migrations/*", "*/tests/*", "*/__pycache__/*"]
branch = true

[tool.coverage.report]
exclude_lines = ["pragma: no cover", "def __repr__", "raise NotImplementedError", "if TYPE_CHECKING:"]
show_missing = true
fail_under = 70

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
